chore(provider): limit forced none thinking to official api
